requirements.txt
google.generativeai
tqdm
gradio
numpy
pandas
scikit-learn
google-genai